Move snp into our shieldbattery dir.
This adds a hook for Storm's snp list initialization process so that we can feed
it a different path to look under, instead of the main Starcraft folder. This
allows us to keep our snp file in our own directory, which should make updates
and other such things way easier.