Create a simple https server.
Includes express setup, basic socket.io setup, and an http server that
redirects to https.  Will require an SSL certificate, I'll add
instructions on how to generate this and make it trusted in various
browsers to the README at a later date.

This commit includes node_modules, I think it makes sense to have these
checked in given that this is a server and not a library.