Fix late CSRF token generation.
When tokens were generated late (e.g. when a user logs in and we
regenerate
their session in the middle of the process), connect was saving the
session
earlier than we were generating the token, leading to sessions without
CSRF
tokens and the client browser having a useless token cookie. This change
manually triggers a save in such cases.

This is not necessarily ideal, since we're now double-saving a session
in short
succession, however I think the cases where we regenerate sessions are
somewhat rare. Thus, this is probably fine.