Add lobby functionality.
Users can view a list of lobbies, join them, and create new lobbies.
Integration with psi is forthcoming, as of now the lobbies don't
actually start a game. There are also many "bugs" surrounding not
closing lobbies at the proper times; these will be cleaned up shortly.

This commit also adds the first set of tests to the repo! The fact
that there aren't more tests is sad :( Soon, or something.