Bring shieldbattery's requirements down to WinXP SP2.
I know somewhere in the bowels of node/libuv, they specify a minimum
of Vista, but on the outside (and in their README) they specify XP SP2.
Things seem to build fine with that requirement (with the exception of
one Registry function that this commit replaces), so we'll go with it.
There may be some other assumptions about the OS level, like in the
injection code, but I'll look around for those later.