Fix build to work on Windows XP.
VS2012's default runtime settings don't support XP, apparently. This
change makes all of our projects get setup with the xp-compatible
runtime, with the downside that we can no longer use static analysis.
I'd like to get static analysis back on at some point, perhaps the
solution is to allow it to be specified as an arg to vcbuild, or to
make XP-specific builds. In any case, I'll do that at a later date.