Make a lobby service so that joined lobbies are persistent.
A joined lobby will now stick with a user as they navigate around the
site, until they either leave it manually or disconnect all their
"active" sockets for that lobby. In a future commit I will track these
sockets more logically, but for now these are any sockets that have
visited the lobby once in their session.

This commit also introduces the first normal "resolve" structure in the
app, along with an attempt to somewhat handle $routeChangeError. Its not
perfect by any means, but its a start I suppose.