Add a user-centric socket wrapper for socket.io.
This adds `io.users`, which is similar to `io.sockets` but allows for
sending messages to all the sockets currently connected for a given
user. We now make use of this in the Lobbies service, where we no
longer have a notion of a socket that is connected for a user in a
lobby but *not* in that lobby (e.g. now no matter whether a tab has
viewed the lobby page or not, it will now be counted as being in that
lobby if the user previously joined it in any other tab).

To add to this, lobby status is now broadcast to users' sockets when
they connect (e.g. if you're in a lobby in another tab, and open a new
tab to the site, that new tab will receive a message telling it that
it is in that lobby). Lobby parts also broadcast to all other sockets
for a user.

One thing I'd still like to do with lobbies is give users a time
period after disconnecting in which they can re-join a lobby, before
we remove them. This would account for cases where a user refreshes
the lobby page, and that is the only SB tab they have open. I think
this should be fairly easy to accomplish after this refactoring, but
I'll leave it for another commit.