Add forge, a windowed mode for the modern age.
Includes basic hooks necessary to stop regular DirectDraw from
initializing, as well as a dummy DirectDraw7 implementation that will
soon have other code to map calls to OpenGL. Forge is meant to be
controlled through JS, thus it is a Node native module, and injects
its hooks based on a call from JS-land.

This commit also includes a change in node-bw to ignore the Wmode
plugin if its requested to be loaded, soon I will remove that and
probably remove plugin loading entirely. Left a TODO there for now.