Create psi-emitter for detecting resolution.
psi-emitter is an executable that takes one argument: a mailslot
path to write resolution info to. When run, it will query the
current resolution and write it to this slot, then exit. If it
fails in any way, it will exit with a non-zero exit code. This allows
us to query such information from a service even though its running
in a headless environment, since psi-emitter can be run under the
active user instead.

Fixes tec27/manner-pylon#8.