Create a script to bundle up all our binaries/scripts.
I'd like to get the JS down into a single file, but it will require a
few changes to _third_party_main.js, so I'll do it in a follow-up
commit. There's also some low-ish hanging fruit to better handle native
modules (which might be nice if/when we include more, e.g. for upnp).