Update _third_party_main to always run the same file.
The main will now run index.js, and pass in the module name. index.js
will decide which other file needs to be required based on this. This
allows us to browserify everything into a single JS file, which A)
simply gives us less files to manager, and B) reduces filesize by not
duplicating a bunch of dependencies.