Create Renderer interface, refactor initialization.
IndirectDraw interfaces now have no idea what graphics library they're
rendering through, and merely ask Forge to create a renderer for them
when the window/resolution settings are set by the game (with the
exception of IndirectDrawPalette, which still has coupling to OpenGL,
but I will fix this in a future commit). This will allow us to write
Renderers for other libraries (e.g. Direct3D) easily and manage swapping
between them with Forge.

This also cleans up the initialization of the OpenGl class in an attempt
to make it not completely convoluted. A factory method was added that
should allow for us to display better errors to users when
initialization fails, although that is not part of this commit. Shader
initialization was partially cleaned up, and I will add a ShaderManager
class in a future CL to further clean that up.