Move to using statuses as well as actions.
This removes what would have been a bunch of repeated boilerplate
for each action, now actions (most, anyway) naturally have a status.

Actions progress through these statuses, and its up to the stores to
check the statuses as necessary. This also updates LOG_OUT to properly
handle failures (although nothing displays those failures as of yet),
and updates the login form to display failures. I *think* this is a
decent way of doing that, and results in the failures being transient
if you browse around the site, while still placing them in the store.