Move WrappedProcess over to nan.
This is the first step in letting us use iojs and replacing node 0.10.
It'll be hard to know whether this migration is completely done until we
try to move to iojs, but thus far its been pretty simple and
straightforward so I imagine we can get it pretty close.

This doesn't include changes to v8-helpers yet, which will be needed, as
I haven't quite figured out where I want to grab a copy of nan.h from for
that library. Will figure this out and fix in a future commit.