Add game result and game length info to the data passed to the runGame callback.
Game results are passed as a map of player name -> game result, and will require
processing to actually be correct (i.e. if a player leaves before the game is
finished, the results that get passed back will still have other players marked
as 'playing' and need to be filled in using later results).

Game time is passed back as milliseconds, assuming fastest game speed.