Begin replacing material-ui with custom stuff.
Not really happy with the direction material-ui is headed in,
namely with all the inline style crap (which clearly has a lot of
problems but they seem to be charging head-first into anyway. Bleh.).
Since we don't actually need a ridiculous number of components and
they don't need to be that flexible, we can probably just use
Polymer/Angular as a reference and implement our own.

I'm attempting to do this without breaking the site completely, so
this commit doesn't actually remove that dep and most of the
components are still material-ui's. The major piece I'm working on
here is the AppBar, which is largely done.

Attempting to match the mocks as I do this, so some strings are
currently set weirdly.