Switch to ESLint (replacing JSHint).
This is a lot more configurable and allows us to lint way more
stuff, as well as supporting more ES6/7/2015/whatever stuff. As
of this commit there's a ton of stuff that's lint erroring, will
fix all of these in follow-up commits.

Lint can be run using one of the integrations (see the eslint site)
or by doing `npm run lint`.