Move to redux instead of facebook flux dispatcher.
Redux looks like it will help simplify our usage here, and just
looks promising in general (especially for hot reloading in dev),
so I think its a good option. Some of this stuff is still sorta
bleeding edge, but feels fairly stable and should stabilize more
in the coming weeks.

Right now only the Auth stuff is working, still need to hook that
up to the siteSocket as well as fix the other things that get
dispatched (server status).