Improve the cursor tracking. Fixes #37.
Before this commit, we would clip the cursor to the BW window as soon as
it activated, regardless of where your cursor was during activation. So
if you alt-tabbed into BW while your cursor was outside BW's window, the
cursor would be automatically moved to the closest point inside BW's
window, which is not intuitive and can be annoying. This commit makes it
so the cursor stays where it's at, and only when you hover over BW window
will it clip the cursor.