Add dummy date for a lobby, lobby in left-nav, adjust styling to account for overflow.
Left nav was expanded slightly (to 3.5x) as it was a bit too
constricing for the length of channel names/lobby names.
Padding was also reduced a bit (matches the mocks better
anyway) for similar reasons. Visuals of the body area were
adjusted to match (and the leftover Card thing from initial
React stuff was removed, that looked awful and bleh).