Move to using CSS modules/postcss instead of stylus.
We lose out on a bit of formatting, as well as some of the more
complicated features of stylus, but we weren't using the vast
majority of them. In exchange, we gain the ability to write CSS
classes that get renamed at runtime (and thus won't have
collision issues) and to be able to reference those via JS
easily, as well as compose in CSS.

Various materials that will be useful with this:
https://github.com/css-modules/css-modules?ts=2
https://github.com/css-modules/postcss-modules-values?ts=2
https://www.npmjs.com/package/postcss-mixins
https://www.npmjs.com/package/postcss-for
http://cssnext.io/features/
https://github.com/jonathantneal/postcss-nesting?ts=2

All of the above are things available in this commit.

During this migration, I noticed/fixed some other things,
including:

- Text sizes are now just specified in px, rem seems not really
ideal/desirable/useful and really makes stuff hard to understand.
See https://mindtheshift.wordpress.com/2015/04/02/r-i-p-rem-viva-css-reference-pixel/

- Shadows were wrong (they were written using an older version
of angular material as reference). They should now be correct
again, and I imagine this will fix issues with the larger depths.

- Removed optimizeLegibility from body text sizes, I imagine
that could (maybe was?) result in bad perf.