Fix issues with interleaving lobby mutations.
Prior to this change, different mutations on the same lobby could
interleave, effectively resulting in cases where we LOAD a lobby,
get pre-empted by a different async "thread" of execution which
then LOADs and STOREs a lobby, then are allowed to execute, mutate
the (now out-of-date) lobby structure and STORE it over the newer
one. This is obviously problematic, and can result in cases where
the state on the client looks completely different than what we
have on the server.

This change fixes this (at least for lobby mutations, ideally we
can also do something similar for e.g. creating/joining a lobby
to lock on a user's lobby state) by appending a Promise to the
end of a chain each time we acquire the lobby for mutating, and
using the body of that promise to perform actual execution.

Note that this sort of code is only necessary when the load and
store of a piece of data occur in separate async blocks. A
synchronous load->mutate->store does not have such issues.