Refactor Select component and add Menu component.
Select and Menu components are similar and they reuse some of the
sub-components which can be found in common/menu-utils. Any element,
or a component can be used to open a Menu when clicked on. If a string is
supplied to the Menu component, it will create an IconButton from that
string and use it as a Menu's element.