Implement working TURBO LOBBIES.
These should be much, much faster than the previous lobby
implementation, at the expense of some work to figure out how
to manually initialize all the stuff we need. The end result is
something that's much, much simpler to deal with given that we
know how the lobby should be configured before the game is
launched.

TODO in future commits:
- Thread a random seed through from the server
- Fix the weird network lag that is "fixed" right now with a
  constantly running interval
- Update state better (and display on site), deal with errors
  better
- Handle other game types/multiple teams