Add maps, thumbnails, and manifest plus an API to list them.
This is a stop-gap to get map selection in with remote
downloading and local caching, without having to evaluate and
test/setup cloud storage systems. Eventually we'll have a system
that will allow for easy user uploads, but not yet.

I wrote a tool to generate the manifest and rename maps and
thumbnails, so if you want to add more maps here, talk to me :)