Don't allow interaction with the site when not connected to it.
In the long run, there's better/smarter things we could do here,
but this is simple and will prevent a lot of problematic things.
We still need to run through the other reducers that depend on
network state and fix them up (e.g. lobby state needs to be
cleared on disconnect).