Improve handling of errors during game launch.
Now errors from the launch are separated out from errors
while waiting for exit, allowing us to properly notify the site
when game launch has failed (which allows for backing out of
initialization for everyone else, hooray!). Error messages should
be much more descriptive thanks to the previous change, and if
the process launches but fails to inject or resume, it will be
terminated instead of being left around.