Implement some fixes for rendering performance/jankiness.
- Make render skipper skip less frames to deal with cases where
we just barely missed the vblank (our goal is to not render frames
that are clearly a waste, not to render every frame we get once).

- Hook a BW function to know when to draw. BW actually calls Lock
one to many times in a single frame, so rendering a new frame on
Unlock is both wasteful, and was resulting in us rendering
partially updated frames to the screen (due to RenderSkipper). Now
we know the full bounds of the screen render, so we just mark
surfaces dirty when they're unlocked, and trigger our actual
drawing once we know the whole screen is done.