Attempt to prevent Windows Games Explorer from injecting StarCraft.
This occurs for some users, and can cause a network call on startup
(before our injection even happens). If that network call fails,
it will be retried until it succeeds (which may be never). Ugh.

We now set an environment flag to try and turn this particular
compatibility shim off (along with some others that would probably
cause issues for us). As a side effect, the environment the SC
process receives is now of the launching user (instead of Psi's).
This is probably what we really wanted anyway, but it means that
ShieldBattery's temp directory (for low-level logs) is now in
%USERPROFILE%\AppData\Local\Temp, not C:\Windows\Temp.