Update node to 6.2.0. [RUN `git submodule sync`]
This makes us use a fork of node for the time being, as we need a
fix applied to one of the ICU gyp files or our build breaks. I'm
working on PRing this into node permanently, but for the time
being I'll maintain a patched version in my own fork.