Rework how game starting works with rally-point in mind.
This sets things up in a single async/await function and runs
through it from start to end (allowing cancellation via a cancel
token class I made). This makes the flow a bit easier to follow,
and should hopefully allow us to pull this out into another class
for matchmaking, etc. easier.

No handling is in yet for actual rally-point usage, so this will
only work for CPU games. Real rally-point support is coming in
the next commit.