Implement the code to read files from user's computer.
And use it to read replays. We impersonate the user when attempting to
read files, because Psi is running under a SYSTEM account and might not
have access to some folders that logged-on user has (eg. Documents
folder).

Note: the function that reads files takes absolute `path` as an argument.
However, proper safeguards should be deployed in the JS-land to make sure
users are not allowed to read arbitrary folders on a user's computer.