Update Popover component to have customizable positioning.
This commit adds four props to the Popover component which control the
positioning of the Popover component:
a) `anchor` - denotes the DOM element on which to base the positioning of
the Popover component.
b) `anchorOffset` - optionally specify an offset for the anchor element on
which the Popover position will be calculated. Accepts an offset on X axis
(horizontal) or on Y axis (vertical). Positive values move position of
anchor down/right; negative values up/left.
c) `anchorOrigin` - specifies the corner on the anchor element on which to
attach the Popover component. Possible options are top/left, top/right,
bottom/left, bottom/right. Default is top/left.
d) `targetOrigin` - specifies the corner on the Popover component which to
use to attach to the `anchorOrigin`. Possible options are top/left,
top/right, bottom/left, bottom/right. Default is top/left.