Add user IP tracking on each request. (#205)
There are 3 cases that we differentiate when a user makes a request:
- This user ID / IP combo doesn't exist in our database -> insert new
  record with first_used and last_used set to current time, counter set to 0
- User ID / IP combo exists and most recent last_used time is more than 1
  hour old -> insert new record with first_used and last_used set to
  current time, increasing the counter by 1
- User ID / IP combo exists and most recent last_used time is less than 1
  hour old -> update the existing record and set last_used to current time

We use a counter for each user ID / IP combo so that we can have unique
key to prevent duplicate insertions, which can often happen if a user has
multiple tabs open or makes multiple requests in very quick succession. In
such cases we expect that DB insertion fails and can be retried.

Part of the implementation for #168.