Get server running again.
Started using yarn for that as well (and removed all
client-only deps from it). There's a bug with yarn right now:
https://github.com/yarnpkg/yarn/issues/2286

That results in it failing to install native deps (like bcrypt).
I have a workaround for this that's simple, but requires
modifying the local copy of node-gyp. That's a one time thing per
yarn install/update. Hopefully yarn will fix it themselves soon
(but for now I'll add the workaround to our setup steps).