Make Process code able to inject DLLs from 64-bit mode.
This adjusts the injection code to find the required functions
(LoadLibraryW, GetProcAddress, and GetLastError) in the target process,
rather than assuming they're the same between processes. I *believe*
this is likely to work better in Wine environments as well, and should
be more resilient to things that hook stuff like GetProcAddress (I know
I used to have problems with ConEmu breaking our injection, that should
be fine now).

I plan to start shipping 64-bit and 32-bit versions of the
standalone app, so this is a necessary thing, and also makes it so
we don't have to set special NPM environment variables to pull down
the right Electron binaries in dev mode (hooray developer experience
:P). A lot of this code was inspired by EasyHook, but has been heavily
adapted for our usage (with proper logged errors and C++ structure
usage).

It'll only work for the 64->32bit direction (well, and 32->32bit), but
that should be fine for our target :)