Create a new local settings manager that can be used over IPC.
This runs in the main process, so that it can save stuff even if
settings get updated just as the renderer window is closing, and so that
it can manage settings around window state. Since it lives there, it
must be operated over IPC (and post updates back there). Settings
related operations have been moved out of nydus routes and into things
that utilize this new IPC stuff.

Settings operations are now merges, rather than direct sets, so that
each side (main process, renderer process, etc.) doesn't have to know
about what the others are setting. The local settings manager now does
fully async operations, so that we can better deal with error cases and
do async initialization on process start.

With this commit, the window state (position, size, and maximized state)
will now be saved and restored on startup. We still need to be good
about possibly changing monitor states so we don't put the window
somewhere inaccessible, but I'll add that in a future commit.