Remove support for using SSL directly in node, fix webpack building for the browser version.
I don't think anyone was using the SSL version anyway, and it makes some
init code more complex. I'd like to make the electron app able to pick
the likely "right" address in dev mode, so getting rid of it simplifies
things a bit.

Webpack brokenness was around pulling in stuff that shouldn't be in the
browser version, looks like webpack either couldn't deal with the early
returns, or wasn't satisfied that the condition we were checking would
always be true.