Make installer creation work.
Had to move the common directory so that it was under the app directory
so that it could get packaged in the ASAR. It's the only thing that
isn't built via webpack, so that seemed like the best solution
(everything else will still be able to pull those files into their
bundles).

This uses an NSIS installer since that seems to be the recommended thing
now, and it builds both architectures into one installer. We can make
this smaller by using nsis-web, but I am waiting until I setup
publishing to start using that. (That installer will go out and download
the proper architecture when it runs.)

In this version, shieldbattery.dll isn't properly signed, I'll add that
in a follow-up commit.

This also adds two separate build configs, one for production and one
for our staging server. They install to separate directories and get
separate user data directories (the local/unpacked dev verison will also
get its own user data directory). Ideally we can maybe change the icon
on the dev one at some point so that we can run them all simultaneously
without being confused, but at the very least they can all be installed
side-by-side now :)