Move to Webpack 2.
This results in a pretty decent size decrease just from the move and
making use of babel-preset-env's useBuiltIns option (240K -> 230K on the
web, and 880K -> 680K on the game JS, not sure about the app JS but
probably similar).

In making this change I noticed that babel was using .babelrc's from the
folders (or ancestor folders) of the JS files being transpiled, which
was unexpected for things like common/. I've changed the repo to only
have 2 babelrc's (one so that it actually gets built into our electron
app, and one at the root, but these files should always match exactly).
For builds that are specific to some client that we don't run from this
directory (e.g. webpack builds), we explicitly disable babelrc usage and
write the options in the webpack configs.

Things feel a bit more performant in the compiled JS too, although I may
just be hallucinating that.