Add throttling for invite signup.
Part of the work for #169.

This uses a token bucket system, so each <id> is given some number of
tokens that refresh every so often, and the ability to burst up to some
number of requests in that time period. The implementation here can be
generalized to a reusable middleware function, but I'll do that in a
near-future commit.