Add an easier way for Forge (c++) to emit events to JS.
This is similar to WrappedBroodWar's EventImmediates, but more centered
around an EventEmitter-like API. Probably this could be pulled out of
Forge and re-used in both places. In this commit, the only thing making
use of these is for window movement (but in theory there is other useful
stuff to know from JS).

Part of #216.