Add a basic icon to the system tray.
Has a single action in the context menu - to quit the app.
This commit also defines the mechanism of exiting/minimizing the app.
Pressing 'X' minimizes the app to the system tray and displays a hint on
how to close the app only the first time the 'X' is clicked.