Add 'unread' overlays to app icons.
Every time there is a new message, we change the app icon and system
tray icon to notify the user that there has been a new message.

Note: we can't use the `setOverlayIcon` API to update the system tray
icon, so we have to replace it with a completely new icon.