Support uncommon UMS computer types
Requires deleting all maps from the db.

The unusual computer types are never displayed in lobby, so had to add
bunch of code to deal with that. The types are just handled as a int,
as while we had string definitions for them, we really don't need to do
anything but pass them to bw.