Implement an AudioManager for loading/playing sounds, use it for join alerts.
This is limited to Electron for now (join alerts are electron-only
anyway), but in the future when we have other sounds (notifications +
clicks and stuff) we'll probably want to make this usable on the web
somehow. I imagine the right way to do that is to split the "loading"
portion from the "playing" portion. Electron would probably continue to
preload all its sounds on startup, the web might try to load/cache them
on demand? Not 100% sure yet, but we'll go with the simpler approach for
now.

This setup has a single gain node feeding into output, we could use that
to increase/decrease the volume of the sounds globally throughout the
app (if we add a setting for it). Some of the ideas behind this setup
come from here:

https://www.html5rocks.com/en/tutorials/webaudio/fieldrunners/

Fixes #311.