Remove invite code requirement from signups.
I left the admin interface intact for now, still need to figure out how
we want to send out an open beta email from that list, as well.

The splash page now looks fairly terrible, but that'll be cleaned up as
we update that anyway. I also removed the card from the signup page as
that was outdated, but it still needs some work to match the other auth
pages.