Refactor game initialization so non-lobby config can be set separately.
This makes it so that we can initialize things like network routes
outside of knowing the game config, which will allow us to pre-emptively
trigger network route setup (under the assumption we can just throw away
routes we don't end up needing).

Currently lobbies are just fixed to use the new commands (but in the
same structure), but they can (and should) be updated to try and
pre-emptively create network routes as people join/leave lobbies (and
completely avoid the loading screen we currently have).