Set up the `prettier` library to aid us in formatting our code.
This sets up the prettier to be a part of our eslint config, so it
shouldn't be necessary to have special scripts to handle prettier on its
own.

Note: We're also using
https://github.com/prettier/eslint-config-prettier so we don't have to
delete all the rules that the prettier conflicts with. There were only a
few rules that had to be manually deleted for some reason.

PS. It might be of interest to developers to add prettier package to
their editor of choice, which allows the programmer to automatically
format the code. Those packages can be found here for each editor:
https://github.com/prettier/prettier#editor-integration