Fix and improve lobby tests.
This commit fixes all the errors in our lobby tests, as well as
refactors some of the tests to be more readable.

There are also a few more tests added to test some of the edge cases
around the lobbies with observers and hidden slots (UMS-related).

As a part of these lobby tests fixes and improvements, there were a
couple of things changed in our lobby code as well:
- fixed a bug when finding the available slot in observer team
- make the presence of observeres in melee lobbies configurable
- make the observer slots 'closed' by default