Copy the local.dll file from StarCraft folder to our downgrade folder. (#371)
* Copy the local.dll file from StarCraft folder to our downgrade folder.

Since 1.19 version, local.dll file is not located in the root StarCraft
folder anymore, but instead in one of the subfolders. To make it easier
for us to use it, we copy it over to our downgrade folder and we'll make
StarCraft look for it there, which will come in a separate PR.

* Hook the LoadLibraryA function to properly load local.dll library.

When BW tries to load the local.dll library, we need to figure out
whether to load it from StarCraft's root directory (if exists), or from
our downgrade folder (for 1.19+ versions).

Fixes #370.