Add and set up styled components.
For now we'll have to keep parallel set ups of both styled components
and css modules, because the move to styled components will be
incremental.

This commit makes color and typography related stuff available to styled
components, with colors being accessible through the `theme` props and
typography stuff through components that can be imported and their style
extended.