Update CSS to workaround the annoying css-modules / postcss bug.
Ever since we upgraded our deps, there is a bug when we try to use
css-module variables inside the `calc` function. For now we have simply
hardcoded the values of those variables, so the site actually works. But
components affected by this bug should be a priority to move over to
styled-components.