Add `userId` and `email` to the email verification link.
Even though we're forcing users to be logged-in before verifying their
emails, which means we could read the above info on the client, doing it
this way ensures the correct email gets verified if the user is
logged-in twice with different accounts, which is possible to do
through two different "environments", ie. an app + browser.