Reset the loading state after the site socket gets disconnected.
This fixes a bug where after navigating to a route which is not below
the SiteConnectedFilter, the site socket would get disconnected, but all
the loading state would remain "loaded". Then after going back to a
route under the SiteConnectedFilter, site socket would get connected
again, but the LoadingFilter would not wait for all the state to be
loaded again since it thought it was already loaded.

This lead the `goToIndex` action creator to incorrectly change the route
based on the current state, which was wrong.

Also, we only reset the loading state of stuff that gets initialized
through the web sockets; other stuff can remain "loaded", eg. an audio
manager.