Make the production build work.
I did run some tests and compared the performance of Webpack 4 on our
project and while the build speed seems quite slower than Webpack 3 (two
times as slow it seems), the final bundle size is almost two times
smaller as well.

Not sure what to think about this. I think it's fine for now, but I
think there's room for additional improvements with code splitting if we
ever need it.