Upgrade Webpack to version 4. (#398)
* Upgrade Webpack to version 4.

Upgrading webpack itself was a lot less painful process than I expected.
However, replacing the hot/dev middleware with new, now suggested,
libraries was a nightmare. Opted out for using an IIFE in server app,
because the `koa-webpack` library for some weird fucking reason returns
a promise when you call it, so you need to wait for it to resolve before
you can use its middleware.

Still requires some testing in the production mode.

* Make the production build work.

I did run some tests and compared the performance of Webpack 4 on our
project and while the build speed seems quite slower than Webpack 3 (two
times as slow it seems), the final bundle size is almost two times
smaller as well.

Not sure what to think about this. I think it's fine for now, but I
think there's room for additional improvements with code splitting if we
ever need it.

* Remove an extra comma.

* Bring back any-promise replacement after all, to eliminate a couple of
webpack warnings