Add support for "One on one" game mode
No idea how it wasn't added since the beginning, but it is here now. We
can use this mode for the 1v1 matchmaking games. Team matchmaking games
will have to use a different mode, but with additional logic (so players
can't change their alliances in the game).

Observing is still limited to melee lobbies only for now, but it
shouldn't be too complicated to support observing in one-on-one lobbies
as well.

Tests haven't been written for this lobby type specifically, since it
pretty much uses the same logic as melee lobbies. And I didn't feel like
writing them :P