Remove the unnecessary explicit cookie applying
Previously this was required because webpack didn't respect the
`browser` field in the `package.json` for the `electron-renderer`
targets, which `engine.io-client` used to determine when to use
which version of `XMLHttpRequest`, so it used the node's version.

Newer version of webpack fixed this, so our explicit setting of
cookies is not required anymore.