Track the last focused state in the File Browser component.
This allows users to continue using the File Browser where they last
left off. The last focused state is saved in the local storage, so as
long as that is not cleared, it should be remembered. This PR also
implements a bunch of shortcuts, which should make using the File
Browser easier, eg. Up Arrow, Down Arrow, Page Up, Page Down, Home and
End.

Note: List virtualization has not been implemented yet, so the
performance in large folders might be questionable. This might come in
the later commit.