Upgrade the `KeyListener` component to handle exclusive listeners.
This commit fixes an undesired behaviour with previous version of the
way the `KeyListener` component worked, which was that all of the
registered handlers would get executed for each key event. A situation
where this behaviour is undesired is when for example a dialog is opened
and you could still type in a chat behind it or even open an activity
overlay through hotkeys.

Now some components that use `KeyListener` (activity overlay, dialog and
popover) are marked as `exclusive`, which means when a keyboard event
happens and there is even one registered listener who is exclusive, it
will only execute handlers for that listener and its descendants. The
reason why we need to also execute handlers for descendants is because
activity overlays, dialogs and popovers, can have descendants who have
key listeners of their own registered and we want to handle those
events.

The way we figure out which key listeners are actually descendants of
the exclusive one might not be an ideal one, but I couldn't come up with
a better one. We basically rely on the order the key listeners are
registered, which sometimes can be a bit unintuitive because of the way
React works and in which order components are mounted.