Add lobby preferences
Preferences load when a user opens the "create lobby" screen and they
update when they close it.
Even though the initial version of the spec document said that we should
calculate "recent maps" based on the maps the user played, I actually
made them into their own thing, because a map can be added there without
actually playing on it (for example if a user selects a local map and
then closes the "create lobby" screen).