Make the game build script runnable from outside the game directory.
There is a script in root folder that builds the game, but the cargo
build fails if it can't find the Cargo.toml file in the directory. We
move to the game folder before running the cargo build, regardless of
the directory we started the script from.

We also remember, and recover the starting directory.

Another little change in this PR is making the *.bat files have CRLF
endings in .editorconfig.