Rework the way the map is selected.
This commit extends the activity overlays functionality to allow them to
be receive some init data. This made it possible to remove some of the
uglier code we had when we had to share the state through the store.

However, this change is not without its drawbacks either. Now create
lobby overlay needs to keep recent maps in its own state since they can
go out of sync with preferences in case a new map is uploaded that
wasn't currently in the recent maps list.

But in the end, this seems much more preferable to what we had before,
especially once we get the server maps browser, which should now work
out of the box with the create lobby overlay.