Fix the map thumbnail in join lobby list.
This is just a quick fix to get the map showing in the join lobby list
with the new map thumbnail component. However, the map name looks
terrible here since it can barely fit on this size of thumbnail.

Gonna leave it as it is for now, as the join lobby screen needs to be
redesigned ASAP anyway.