Upgrade the lobby creation workflow, both in functionality and in design. (#406)
* Refactor the server-side of map distribution code

This is mostly done in preparation for the upcoming changes to the map
distribution code which allowed us to simplify some of the things. Still
need to refactor the map parsing worker a bit and fix any bugs that I
have missed in this refactor.

* Refactor the client-side of map distribution code

This should pretty much make everything work again, just with a lot less
code which is always a plus!

* Seperate the logic for server-side maps and local maps a bit more clearly

* Add lobby preferences

Preferences load when a user opens the "create lobby" screen and they
update when they close it.
Even though the initial version of the spec document said that we should
calculate "recent maps" based on the maps the user played, I actually
made them into their own thing, because a map can be added there without
actually playing on it (for example if a user selects a local map and
then closes the "create lobby" screen).

* Start updating the create lobby design to match new mocks

* Make the logic of selecting a local map a bit more explicit and make
shit actually work

* Preserve the order of the input array when getting map infos.

This was actually present in the original code, but I mistakenly took it
out. Oops.

* Always mark the selected map in the list of recent maps, when uploading
a map

* Fix the config for react-svg-loader to not strip the 'viewBox' attribute

* Make the return type of 'getMapInfo' consistent.

* Send the meta properties of 'ACTIVITY_OVERLAY_CLOSE' action as an object.

* Use a correct function to check game sub type.

* Actually wait before files are written on disk before returning info.

* Update activity overlay background color to match our new color scheme.

* Bring the design of lobby creation to match the mocks.

NOTE: Still missing the screen to select the maps from the server, but
that will come in a later commit, possible even a separate PR.

* Use the zoom-crop for thumbnail images instead of stretching them.

* Add focus trap to activity overlays.

* Add ability to cycle through maps with TAB and select maps with SPACE.

* Fix a bug in the Select component by not handling key presses if the
Select wasn't focused.

By pressing Enter on any other component while the Select component was
rendered, there was potential for Select component to hijack the Enter
key, even while it was not focused.

* Add validation for map selection.

* Add the back button to the browse-local-maps overlay.

* Handle the case when the selected map is invalid.

* Allow MapSelect component to be focused only with the keyboard.

* Submit the create-lobby form by pressing an Enter key.

Also moved the Select event handlers from `keypress` event to `keydown`
event (since `keypress` event has been deprecated and create-lobby needs
to use the same event so it doesn't overwrite Select's handling of the
Enter key).

* Unfocus the currently focused map on a mouse click.

* Fix a bug where the selected map would be wrong after hosting a game.

* Bump the minimum required version of Postgres in getting started docs.

* Move the selected map to the front of the recent maps list when hosted.

* Don't use state for  flag.

* Rework the way the map is selected.

This commit extends the activity overlays functionality to allow them to
be receive some init data. This made it possible to remove some of the
uglier code we had when we had to share the state through the store.

However, this change is not without its drawbacks either. Now create
lobby overlay needs to keep recent maps in its own state since they can
go out of sync with preferences in case a new map is uploaded that
wasn't currently in the recent maps list.

But in the end, this seems much more preferable to what we had before,
especially once we get the server maps browser, which should now work
out of the box with the create lobby overlay.

* Turn on the feature flag for uploading maps. Wohoo!

* On second thought, keep the map upload feature dev-only for now.

Any map that is uploaded before we introduce map ownership concept
(which there is a PR for) is pretty much useless and will be discarded.

* Minor cleanup

* Fix the map thumbnail in join lobby list.

This is just a quick fix to get the map showing in the join lobby list
with the new map thumbnail component. However, the map name looks
terrible here since it can barely fit on this size of thumbnail.

Gonna leave it as it is for now, as the join lobby screen needs to be
redesigned ASAP anyway.

* Show the full map name in the tooltip when hovered over map name.